Extract DataGridView nested classes
