+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridView.DataGridViewValidateCellInternal.cs
@@ -1,0 +1,18 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+#nullable disable
+
+namespace System.Windows.Forms
+{
+    public partial class DataGridView
+    {
+        internal enum DataGridViewValidateCellInternal
+        {
+            Never,
+            Always,
+            WhenChanged
+        }
+    }
+}
