+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridView.cs
@@ -5470,76 +5470,5 @@ namespace System.Windows.Forms
                 }
             }
         }
-
-        /* INTERNAL ENUMERATIONS */
-
-        internal enum DataGridViewHitTestTypeInternal
-        {
-            None,
-            Cell,
-            ColumnHeader,
-            RowHeader,
-            ColumnResizeLeft,
-            ColumnResizeRight,
-            RowResizeTop,
-            RowResizeBottom,
-            FirstColumnHeaderLeft,
-            TopLeftHeader,
-            TopLeftHeaderResizeLeft,
-            TopLeftHeaderResizeRight,
-            TopLeftHeaderResizeTop,
-            TopLeftHeaderResizeBottom,
-            ColumnHeadersResizeBottom,
-            ColumnHeadersResizeTop,
-            RowHeadersResizeRight,
-            RowHeadersResizeLeft,
-            ColumnHeaderLeft,
-            ColumnHeaderRight
-        }
-
-        internal enum DataGridViewValidateCellInternal
-        {
-            Never,
-            Always,
-            WhenChanged
-        }
-
-        private enum DataGridViewMouseEvent
-        {
-            Click,
-            DoubleClick,
-            MouseClick,
-            MouseDoubleClick,
-            MouseDown,
-            MouseUp,
-            MouseMove
-        }
-
-        private struct MouseClickInfo
-        {
-            public MouseButtons button;
-            public long timeStamp;
-            public int x;
-            public int y;
-            public int col;
-            public int row;
-        }
-
-        internal class DataGridViewEditingPanel : Panel
-        {
-            private readonly DataGridView owningDataGridView;
-
-            public DataGridViewEditingPanel(DataGridView owningDataGridView)
-            {
-                this.owningDataGridView = owningDataGridView;
-            }
-
-            internal override bool SupportsUiaProviders => true;
-
-            protected override AccessibleObject CreateAccessibilityInstance()
-            {
-                return new DataGridViewEditingPanelAccessibleObject(owningDataGridView, this);
-            }
-        }
     }
 }
