+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewColumnHeaderCell.DataGridViewColumnHeaderCellRenderer.cs
@@ -1,0 +1,53 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+#nullable disable
+
+using System.Drawing;
+using System.Windows.Forms.VisualStyles;
+
+namespace System.Windows.Forms
+{
+    public partial class DataGridViewColumnHeaderCell
+    {
+        private class DataGridViewColumnHeaderCellRenderer
+        {
+            private static VisualStyleRenderer visualStyleRenderer;
+
+            private DataGridViewColumnHeaderCellRenderer()
+            {
+            }
+
+            public static VisualStyleRenderer VisualStyleRenderer
+            {
+                get
+                {
+                    if (visualStyleRenderer == null)
+                    {
+                        visualStyleRenderer = new VisualStyleRenderer(HeaderElement);
+                    }
+                    return visualStyleRenderer;
+                }
+            }
+
+            public static void DrawHeader(Graphics g, Rectangle bounds, int headerState)
+            {
+                Rectangle rectClip = Rectangle.Truncate(g.ClipBounds);
+                if ((int)HeaderItemState.Hot == headerState)
+                {
+                    // Workaround for a
+                    VisualStyleRenderer.SetParameters(HeaderElement);
+                    Rectangle cornerClip = new Rectangle(bounds.Left, bounds.Bottom - 2, 2, 2);
+                    cornerClip.Intersect(rectClip);
+                    VisualStyleRenderer.DrawBackground(g, bounds, cornerClip);
+                    cornerClip = new Rectangle(bounds.Right - 2, bounds.Bottom - 2, 2, 2);
+                    cornerClip.Intersect(rectClip);
+                    VisualStyleRenderer.DrawBackground(g, bounds, cornerClip);
+                }
+                VisualStyleRenderer.SetParameters(HeaderElement.ClassName, HeaderElement.Part, headerState);
+                VisualStyleRenderer.DrawBackground(g, bounds, rectClip);
+            }
+        }
+    }
+}
