+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewComboBoxCell.cs
@@ -15,7 +15,7 @@ using static Interop;
 
 namespace System.Windows.Forms
 {
-    public class DataGridViewComboBoxCell : DataGridViewCell
+    public partial class DataGridViewComboBoxCell : DataGridViewCell
     {
         private static readonly int PropComboBoxCellDataSource = PropertyStore.CreateKey();
         private static readonly int PropComboBoxCellDisplayMember = PropertyStore.CreateKey();
@@ -2955,142 +2955,6 @@ namespace System.Windows.Forms
 
                 CompareInfo compInfo = Application.CurrentCulture.CompareInfo;
                 return compInfo.Compare(itemName1, itemName2, CompareOptions.StringSort);
-            }
-        }
-
-        private class DataGridViewComboBoxCellRenderer
-        {
-            [ThreadStatic]
-            private static VisualStyleRenderer visualStyleRenderer;
-            private static readonly VisualStyleElement ComboBoxBorder = VisualStyleElement.ComboBox.Border.Normal;
-            private static readonly VisualStyleElement ComboBoxDropDownButtonRight = VisualStyleElement.ComboBox.DropDownButtonRight.Normal;
-            private static readonly VisualStyleElement ComboBoxDropDownButtonLeft = VisualStyleElement.ComboBox.DropDownButtonLeft.Normal;
-            private static readonly VisualStyleElement ComboBoxReadOnlyButton = VisualStyleElement.ComboBox.ReadOnlyButton.Normal;
-
-            private DataGridViewComboBoxCellRenderer()
-            {
-            }
-
-            public static VisualStyleRenderer VisualStyleRenderer
-            {
-                get
-                {
-                    if (visualStyleRenderer == null)
-                    {
-                        visualStyleRenderer = new VisualStyleRenderer(ComboBoxReadOnlyButton);
-                    }
-                    return visualStyleRenderer;
-                }
-            }
-
-            public static void DrawTextBox(Graphics g, Rectangle bounds, ComboBoxState state)
-            {
-                ComboBoxRenderer.DrawTextBox(g, bounds, state);
-            }
-
-            public static void DrawDropDownButton(Graphics g, Rectangle bounds, ComboBoxState state)
-            {
-                ComboBoxRenderer.DrawDropDownButton(g, bounds, state);
-            }
-
-            // Post theming functions
-            public static void DrawBorder(Graphics g, Rectangle bounds)
-            {
-                if (visualStyleRenderer == null)
-                {
-                    visualStyleRenderer = new VisualStyleRenderer(ComboBoxBorder);
-                }
-                else
-                {
-                    visualStyleRenderer.SetParameters(ComboBoxBorder.ClassName, ComboBoxBorder.Part, ComboBoxBorder.State);
-                }
-                visualStyleRenderer.DrawBackground(g, bounds);
-            }
-
-            public static void DrawDropDownButton(Graphics g, Rectangle bounds, ComboBoxState state, bool rightToLeft)
-            {
-                if (rightToLeft)
-                {
-                    if (visualStyleRenderer == null)
-                    {
-                        visualStyleRenderer = new VisualStyleRenderer(ComboBoxDropDownButtonLeft.ClassName, ComboBoxDropDownButtonLeft.Part, (int)state);
-                    }
-                    else
-                    {
-                        visualStyleRenderer.SetParameters(ComboBoxDropDownButtonLeft.ClassName, ComboBoxDropDownButtonLeft.Part, (int)state);
-                    }
-                }
-                else
-                {
-                    if (visualStyleRenderer == null)
-                    {
-                        visualStyleRenderer = new VisualStyleRenderer(ComboBoxDropDownButtonRight.ClassName, ComboBoxDropDownButtonRight.Part, (int)state);
-                    }
-                    else
-                    {
-                        visualStyleRenderer.SetParameters(ComboBoxDropDownButtonRight.ClassName, ComboBoxDropDownButtonRight.Part, (int)state);
-                    }
-                }
-                visualStyleRenderer.DrawBackground(g, bounds);
-            }
-
-            public static void DrawReadOnlyButton(Graphics g, Rectangle bounds, ComboBoxState state)
-            {
-                if (visualStyleRenderer == null)
-                {
-                    visualStyleRenderer = new VisualStyleRenderer(ComboBoxReadOnlyButton.ClassName, ComboBoxReadOnlyButton.Part, (int)state);
-                }
-                else
-                {
-                    visualStyleRenderer.SetParameters(ComboBoxReadOnlyButton.ClassName, ComboBoxReadOnlyButton.Part, (int)state);
-                }
-                visualStyleRenderer.DrawBackground(g, bounds);
-            }
-        }
-
-        protected class DataGridViewComboBoxCellAccessibleObject : DataGridViewCellAccessibleObject
-        {
-            public DataGridViewComboBoxCellAccessibleObject(DataGridViewCell owner) : base(owner)
-            {
-            }
-
-            internal override bool IsIAccessibleExSupported() => true;
-
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                switch (propertyID)
-                {
-                    case UiaCore.UIA.ControlTypePropertyId:
-                        return UiaCore.UIA.ComboBoxControlTypeId;
-                    case UiaCore.UIA.IsExpandCollapsePatternAvailablePropertyId:
-                        return IsPatternSupported(UiaCore.UIA.ExpandCollapsePatternId);
-                }
-
-                return base.GetPropertyValue(propertyID);
-            }
-
-            internal override bool IsPatternSupported(UiaCore.UIA patternId)
-            {
-                if (patternId == UiaCore.UIA.ExpandCollapsePatternId)
-                {
-                    return true;
-                }
-
-                return base.IsPatternSupported(patternId);
-            }
-
-            internal override UiaCore.ExpandCollapseState ExpandCollapseState
-            {
-                get
-                {
-                    DataGridViewComboBoxEditingControl comboBox = Owner.Properties.GetObject(PropComboBoxCellEditingComboBox) as DataGridViewComboBoxEditingControl;
-                    if (comboBox != null)
-                    {
-                        return comboBox.DroppedDown ? UiaCore.ExpandCollapseState.Expanded : UiaCore.ExpandCollapseState.Collapsed;
-                    }
-
-                    return UiaCore.ExpandCollapseState.Collapsed;
-                }
             }
         }
     }
