+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewLinkCell.cs
@@ -13,7 +13,7 @@ using static Interop;
 
 namespace System.Windows.Forms
 {
-    public class DataGridViewLinkCell : DataGridViewCell
+    public partial class DataGridViewLinkCell : DataGridViewCell
     {
         private static readonly DataGridViewContentAlignment anyLeft = DataGridViewContentAlignment.TopLeft | DataGridViewContentAlignment.MiddleLeft | DataGridViewContentAlignment.BottomLeft;
         private static readonly DataGridViewContentAlignment anyRight = DataGridViewContentAlignment.TopRight | DataGridViewContentAlignment.MiddleRight | DataGridViewContentAlignment.BottomRight;
@@ -1120,54 +1120,6 @@ namespace System.Windows.Forms
         public override string ToString()
         {
             return "DataGridViewLinkCell { ColumnIndex=" + ColumnIndex.ToString(CultureInfo.CurrentCulture) + ", RowIndex=" + RowIndex.ToString(CultureInfo.CurrentCulture) + " }";
-        }
-
-        protected class DataGridViewLinkCellAccessibleObject : DataGridViewCellAccessibleObject
-        {
-            public DataGridViewLinkCellAccessibleObject(DataGridViewCell owner) : base(owner)
-            {
-            }
-
-            public override string DefaultAction
-            {
-                get
-                {
-                    return SR.DataGridView_AccLinkCellDefaultAction;
-                }
-            }
-
-            public override void DoDefaultAction()
-            {
-                DataGridViewLinkCell dataGridViewCell = (DataGridViewLinkCell)Owner;
-                DataGridView dataGridView = dataGridViewCell.DataGridView;
-
-                if (dataGridView != null && dataGridViewCell.RowIndex == -1)
-                {
-                    throw new InvalidOperationException(SR.DataGridView_InvalidOperationOnSharedCell);
-                }
-
-                if (dataGridViewCell.OwningColumn != null && dataGridViewCell.OwningRow != null)
-                {
-                    dataGridView.OnCellContentClickInternal(new DataGridViewCellEventArgs(dataGridViewCell.ColumnIndex, dataGridViewCell.RowIndex));
-                }
-            }
-
-            public override int GetChildCount()
-            {
-                return 0;
-            }
-
-            internal override bool IsIAccessibleExSupported() => true;
-
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                if (propertyID == UiaCore.UIA.ControlTypePropertyId)
-                {
-                    return UiaCore.UIA.HyperlinkControlTypeId;
-                }
-
-                return base.GetPropertyValue(propertyID);
-            }
         }
     }
 }
