+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewRowHeaderCell.DataGridViewRowHeaderCellRenderer.cs
@@ -1,0 +1,42 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+#nullable disable
+
+using System.Drawing;
+using System.Windows.Forms.VisualStyles;
+
+namespace System.Windows.Forms
+{
+    public partial class DataGridViewRowHeaderCell
+    {
+        private class DataGridViewRowHeaderCellRenderer
+        {
+            private static VisualStyleRenderer visualStyleRenderer;
+
+            private DataGridViewRowHeaderCellRenderer()
+            {
+            }
+
+            public static VisualStyleRenderer VisualStyleRenderer
+            {
+                get
+                {
+                    if (visualStyleRenderer == null)
+                    {
+                        visualStyleRenderer = new VisualStyleRenderer(HeaderElement);
+                    }
+
+                    return visualStyleRenderer;
+                }
+            }
+
+            public static void DrawHeader(Graphics g, Rectangle bounds, int headerState)
+            {
+                VisualStyleRenderer.SetParameters(HeaderElement.ClassName, HeaderElement.Part, headerState);
+                VisualStyleRenderer.DrawBackground(g, bounds, Rectangle.Truncate(g.ClipBounds));
+            }
+        }
+    }
+}
